Made progress on pastel impl
