+++ src/frontend/mod.rs
@@ -1,4 +1,3 @@
-pub mod colors;
 pub mod errors;
 pub mod lexer;
 pub mod source;
