+++ src/main.rs
@@ -1,20 +1,11 @@
 mod frontend;
+mod pastel;
 
-use frontend::colors::ColorOption;
+use pastel::pastel;
 
 fn main() {
-    let col = ColorOption::bold_and(Some("l:white".to_string()), Some("l:magenta".to_string()));
-
-    println!(
-        "{}",
-        col.format_text("Bold white text on magenta background!")
-    );
-
-    let underline =
-        ColorOption::underline_and(Some("d:black".to_string()), Some("d:white".to_string()));
-
-    println!(
-        "\n{}",
-        underline.format_text("Underlined black text on white background!")
-    );
+    let text = r"
+    
+    ";
+    println!("{}", pastel(text));
 }
